+++ map_gpt_embeddings/tap.py
@@ -1,0 +1,8 @@
+from singer_sdk.tap_base import Tap
+
+class TapOpenAI(Tap):
+    """OpenAI tap class."""
+
+    name = "tap-openai"
+    def discover_streams(self):
+        return []
